+++ pkgs/test/lib/src/bootstrap/node.dart
@@ -2,17 +2,19 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../runner/node/socket_channel.dart';
-
+import 'package:test_api/src/backend/stack_trace_formatter.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/plugin/remote_platform_helpers.dart'; // ignore: implementation_imports
-import 'package:test_core/src/util/stack_trace_mapper.dart'; // ignore: implementation_imports
+
+import '../runner/node/socket_channel.dart';
+import '../util/stack_trace_mapper.dart';
 
 /// Bootstraps a browser test to communicate with the test runner.
 void internalBootstrapNodeTest(Function Function() getMain) {
   var channel = serializeSuite(getMain, beforeLoad: () async {
     var serialized = await suiteChannel('test.node.mapper').stream.first;
     if (serialized == null || serialized is! Map) return;
-    setStackTraceMapper(JSStackTraceMapper.deserialize(serialized as Map));
+    StackTraceFormatter.current.configure(
+        mapper: JSStackTraceMapper.deserialize(serialized as Map).mapStackTrace);
   });
   socketChannel().pipe(channel);
 }