+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.12.1-dev
+
 ## 1.12.0
 
 * Bump minimum SDK to `2.4.0` for safer usage of for-loop elements.